// types/types.ts
import { Process } from "framer-motion";
import { EnvVars } from "next/dist/build/turborepo-access-trace/types";

// types.ts - Định nghĩa các interface

export interface WooCommerceConfig {
  baseUrl: string;
  consumerKey: string;
  consumerSecret: string;
  version?: string;
}


export interface WooCommerceProduct {
  id: number;
  name: string;
  slug: string;
  permalink: string;
  date_created: string;
  date_modified: string;
  type: string;
  status: string;
  featured: boolean;
  catalog_visibility: string;
  description: string;
  short_description: string;
  sku: string;
  price: string;
  regular_price: string;
  sale_price: string;
  on_sale: boolean;
  purchasable: boolean;
  total_sales: number;
  virtual: boolean;
  downloadable: boolean;
  downloads: any[];
  download_limit: number;
  download_expiry: number;
  external_url: string;
  button_text: string;
  tax_status: string;
  tax_class: string;
  manage_stock: boolean;
  stock_quantity: number;
  stock_status: string;
  backorders: string;
  backorders_allowed: boolean;
  backordered: boolean;
  sold_individually: boolean;
  weight: string;
  dimensions: {
    length: string;
    width: string;
    height: string;
  };
  shipping_required: boolean;
  shipping_taxable: boolean;
  shipping_class: string;
  shipping_class_id: number;
  reviews_allowed: boolean;
  average_rating: string;
  rating_count: number;
  related_ids: number[];
  upsell_ids: number[];
  cross_sell_ids: number[];
  parent_id: number;
  purchase_note: string;
  categories: Array<{
    id: number;
    name: string;
    slug: string;
  }>;
  tags: Array<{
    id: number;
    name: string;
    slug: string;
  }>;
  images: Array<{
    id: number;
    date_created: string;
    date_modified: string;
    src: string;
    name: string;
    alt: string;
  }>;
  attributes: Array<{
    id: number;
    name: string;
    position: number;
    visible: boolean;
    variation: boolean;
    options: string[];
  }>;
  default_attributes: any[];
  variations: number[];
  grouped_products: number[];
  menu_order: number;
}


export interface ProductFilters {
  search?: string;
  category?: number;
  tag?: number;
  featured?: boolean;
  on_sale?: boolean;
  min_price?: number;
  max_price?: number;
  status?: 'draft' | 'pending' | 'private' | 'publish';
  orderby?: 'date' | 'id' | 'include' | 'title' | 'slug' | 'price' | 'popularity' | 'rating' | 'menu_order';
  order?: 'asc' | 'desc';
  per_page?: number;
  page?: number;
}

// Category

export interface WooProductCategory {
  id:       number;
  name:	   string;
  slug:	    string;
  parent:	      number;
  description:	string;
  display:	  string;
  image:    	object;
  menu_order:  number;
  count: number;
}

export interface CategoryFilter {
  per_page?: number;
  page?: number;
  id?: number;
  name?: string;
  slug?: string;
  parent?: number;
  description?: string;
}